+++ scripts/lib/utilities/crosshairUtils.js
@@ -1,6 +1,5 @@
 import {Crosshairs} from '../crosshairs.js';
-import {genericUtils} from './genericUtils.js';
-
+import {genericUtils} from '../../utils.js';
 async function aimCrosshair({token, maxRange, crosshairsConfig, centerpoint, drawBoundries, customCallbacks, trackDistance = true, fudgeDistance = 0, validityFunctions = []}) {
     let distance = 0;
     let widthAdjust = 0;