+++ scripts/macros.js
@@ -1,34 +1,34 @@
 // Generic Monster Features
+export {abilityDrain} from './macros/2014/monsterFeatures/generic/abilityDrain.js';
 export {advantageDamageBonus} from './macros/2014/monsterFeatures/generic/advantageDamageBonus.js';
 export {auraDamageEnd} from './macros/2014/monsterFeatures/generic/auraDamage.js';
 export {autoGrapple} from './macros/2014/monsterFeatures/generic/autoGrapple.js';
+export {autoPush} from './macros/2014/monsterFeatures/generic/autoPush.js';
 export {berserk} from './macros/2014/monsterFeatures/generic/berserk.js';
 export {contextualBonus} from './macros/2014/monsterFeatures/generic/contextualBonus.js';
 export {damageTurnStart} from './macros/2014/monsterFeatures/generic/damageTurnStart.js';
 export {deathBurst} from './macros/2014/monsterFeatures/generic/deathBurst.js';
+export {effectImmunity, effectImmunityRemove} from './macros/2014/monsterFeatures/generic/effectImmunity.js';
 export {enlarge, enlargeEnlarged} from './macros/2014/monsterFeatures/generic/enlarge.js';
 export {gaze} from './macros/2014/monsterFeatures/generic/gaze.js';
 export {genericRegeneration} from './macros/2014/monsterFeatures/generic/regeneration.js';
 export {genericTeleport} from './macros/2014/monsterFeatures/generic/teleport.js';
 export {keenSenses} from './macros/2014/monsterFeatures/generic/keenSenses.js';
+export {lifesteal} from './macros/2014/monsterFeatures/generic/lifesteal.js';
+export {martialAdvantage} from './macros/2014/monsterFeatures/generic/martialAdvantage.js';
 export {packTactics} from './macros/2014/monsterFeatures/generic/packTactics.js';
+export {parry} from './macros/2014/monsterFeatures/generic/parry.js';
 export {reduce, reduceReduced} from './macros/2014/monsterFeatures/generic/reduce.js';
 export {reduceMaxHP} from './macros/2014/monsterFeatures/generic/reduceMaxHP.js';
 export {spellTurning} from './macros/2014/monsterFeatures/generic/spellTurning.js';
-export {abilityDrain} from './macros/2014/monsterFeatures/generic/abilityDrain.js';
 export {sunlightSensitivity} from './macros/2014/monsterFeatures/generic/sunlightSensitivity.js';
 export {surpriseAttack} from './macros/2014/monsterFeatures/generic/surpriseAttack.js';
 export {swarmDamage} from './macros/2014/monsterFeatures/generic/swarmDamage.js';
 export {touchDamage} from './macros/2014/monsterFeatures/generic/touchDamage.js';
 export {undeadFortitude} from './macros/2014/monsterFeatures/generic/undeadFortitude.js';
-export {martialAdvantage} from './macros/2014/monsterFeatures/generic/martialAdvantage.js';
-export {parry} from './macros/2014/monsterFeatures/generic/parry.js';
-export {lifesteal} from './macros/2014/monsterFeatures/generic/lifesteal.js';
-export {effectImmunity, effectImmunityRemove} from './macros/2014/monsterFeatures/generic/effectImmunity.js';
-export {autoPush} from './macros/2014/monsterFeatures/generic/autoPush.js';
 //Mechanics
-export {invisible} from './macros/2024/mechanics/invisible.js';
 export {cleaveMastery, grazeMastery, pushMastery, sapMastery, slowMastery, toppleMastery, vexMastery} from './macros/2024/mechanics/masteries.js';
+export {invisible} from './macros/2024/mechanics/invisible.js';
 //Generic
 export {proneOnFail} from './macros/2024/generic/proneOnFail.js';
 //Spells
@@ -119,8 +119,8 @@ export {protectionFromEvilAndGood} from './macros/2024/spells/protectionFromEvil
 export {sanctuary, sanctuarySafe} from './macros/2024/spells/sanctuary.js';
 export {scorchingRay} from './macros/2024/spells/scorchingRay.js';
 export {searingSmite} from './macros/2024/spells/searingSmite.js';
-export {shield} from './macros/2024/spells/shield.js';
 export {shieldOfFaith} from './macros/2024/spells/shieldOfFaith.js';
+export {shield} from './macros/2024/spells/shield.js';
 export {shiningSmite} from './macros/2024/spells/shiningSmite.js';
 export {shockingGrasp} from './macros/2024/spells/shockingGrasp.js';
 export {sleetStorm, sleetStormArea} from './macros/2024/spells/sleetStorm.js';
@@ -152,28 +152,28 @@ export {wish, wishStress, wishImmunity} from './macros/2024/spells/wish.js';
 export {wrathfulSmite} from './macros/2024/spells/wrathfulSmite.js';
 export {zoneOfTruth} from './macros/2024/spells/zoneOfTruth.js';
 //Items
-export {unarmedStrike} from './macros/2024/items/misc/unarmedStrike.js';
 export {healersKit} from './macros/2024/items/healersKit.js';
+export {unarmedStrike} from './macros/2024/items/misc/unarmedStrike.js';
 //Feats
 export {alert} from './macros/2024/feats/alert.js';
+export {archery} from './macros/2024/feats/archery.js';
 export {crusher} from './macros/2024/feats/crusher.js';
+export {defense} from './macros/2024/feats/defense.js';
+export {dueling} from './macros/2024/feats/dueling.js';
 export {elementalAdeptA, elementalAdeptC, elementalAdeptF, elementalAdeptL, elementalAdeptT} from './macros/2024/feats/elementalAdept.js';
 export {grappler} from './macros/2024/feats/grappler.js';
+export {greatWeaponFighting} from './macros/2024/feats/greatWeaponFighting.js';
 export {healer} from './macros/2024/feats/healer.js';
 export {heavyArmorMaster} from './macros/2024/feats/heavyArmorMaster.js';
+export {interception} from './macros/2024/feats/interception.js';
 export {piercer} from './macros/2024/feats/piercer.js';
+export {protection, protectionMoved} from './macros/2024/feats/protection.js';
 export {savageAttacker} from './macros/2024/feats/savageAttacker.js';
 export {sharpshooter} from './macros/2024/feats/sharpshooter.js';
 export {slasher} from './macros/2024/feats/slasher.js';
 export {speedy} from './macros/2024/feats/speedy.js';
 export {tavernBrawler, tavernBrawlerUnarmedStrike} from './macros/2024/feats/tavernBrawler.js';
 export {telekineticShove} from './macros/2024/feats/telekinetic.js';
-export {archery} from './macros/2024/feats/archery.js';
-export {defense} from './macros/2024/feats/defense.js';
-export {dueling} from './macros/2024/feats/dueling.js';
-export {greatWeaponFighting} from './macros/2024/feats/greatWeaponFighting.js';
-export {interception} from './macros/2024/feats/interception.js';
-export {protection, protectionMoved} from './macros/2024/feats/protection.js';
 export {thrownWeaponFighting} from './macros/2024/feats/thrownWeaponFighting.js';
 export {unarmedFighting, unarmedFightingUnarmedStrike} from './macros/2024/feats/unarmedFighting.js';
 //Class Features
@@ -244,3 +244,10 @@ export {greaterDivineIntervention, greaterDivineInterventionRest} from './macros
 export {discipleOfLife} from './macros/2024/classFeatures/cleric/life/discipleOfLife.js';
 export {preserveLife} from './macros/2024/classFeatures/cleric/life/preserveLife.js';
 export {blessedHealer} from './macros/2024/classFeatures/cleric/life/blessedHealer.js';
+//Light Cleric
+export {radianceOfTheDawn} from './macros/2024/classFeatures/cleric/light/radianceOfTheDawn.js';
+export {wardingFlare} from './macros/2024/classFeatures/cleric/light/wardingFlare.js';
+export {improvedWardingFlare} from './macros/2024/classFeatures/cleric/light/improvedWardingFlare.js';
+export {coronaOfLight, coronaOfLightEffect} from './macros/2024/classFeatures/cleric/light/coronaOfLight.js';
+// Trickery Domain
+export {blessingOfTheTrickster} from './macros/2024/classFeatures/cleric/trickery/blessingOfTheTrickster.js';