+++ scripts/macros/2014/raceFeatures/loxodon/loxodonSerenity.js
@@ -1,0 +1,5 @@
+export let loxodonSerenity = {
+    name: 'Loxodon Serenity',
+    version: '1.3.10',
+    rules: 'legacy'
+};