+++ scripts/macros/2024/classFeatures/bard/dance/tandemFootwork.js
@@ -1,8 +1,20 @@
-import {activityUtils, genericUtils, itemUtils} from '../../../../../utils.js';
+import {activityUtils, effectUtils, genericUtils, itemUtils, rollUtils} from '../../../../../utils.js';
 import {bardicInspiration} from '../bardicInspiration.js';
 async function early({trigger, workflow}) {
     if (!workflow.token) return;
     await genericUtils.updateTargets(Array.from(workflow.targets).concat(workflow.token));
+}
+async function use({trigger, workflow}) {
+    let classIdentifier = itemUtils.getConfig(workflow.item, 'classIdentifier');
+    let scaleIdentifier = itemUtils.getConfig(workflow.item, 'scaleIdentifier');
+    let formula = workflow.actor.system.scale[classIdentifier][scaleIdentifier].formula;
+    let roll = await rollUtils.rollDice(formula, {chatMessage: true});
+    let sourceEffect = workflow.item.effects.contents?.[0];
+    if (!sourceEffect) return;
+    let effectData = genericUtils.duplicate(sourceEffect.toObject());
+    effectData.duration = itemUtils.convertDuration(workflow.activity);
+    effectData.changes[0].value = roll.roll.total;
+    await Promise.all(workflow.targets.map(async token => await effectUtils.createEffect(token.actor, effectData)));
 }
 async function added({trigger: {entity: item, identifier, actor}}) {
     let bardicInspiration = itemUtils.getItemByIdentifier(actor, 'bardicInspiration');
@@ -30,7 +42,7 @@ async function updateScales(origItem, newItemData) {
 }
 export let tandemFootwork = {
     name: 'Tandem Footwork',
-    version: '1.1.36',
+    version: '1.3.8',
     rules: 'modern',
     early: updateScales,
     midi: {
@@ -38,6 +50,11 @@ export let tandemFootwork = {
             {
                 pass: 'preambleComplete',
                 macro: early,
+                priority: 50
+            },
+            {
+                pass: 'rollFinished',
+                macro: use,
                 priority: 50
             }
         ]